+++ character/admin.py
@@ -2,13 +2,13 @@ from django.contrib import admin
 
 from .models.advancement import Advancement
 from .models.character import Character
-from .models.classes import ClassAdvancement, HitPoints
+from .models.klasses import KlassAdvancement, HitPoints
 from .models.races import Sense
 
 
 class CharacterAdmin(admin.ModelAdmin):
-    fields = ["name", "user", "race", "class_name"]
-    list_display = ["name", "user", "race", "class_name", "level", "xp"]
+    fields = ["name", "user", "race", "klass"]
+    list_display = ["name", "user", "race", "klass", "level", "xp"]
 
 
 class AdvancementAdmin(admin.ModelAdmin):
@@ -22,14 +22,14 @@ class SenseAdmin(admin.ModelAdmin):
 
 
 class HitPointsAdmin(admin.ModelAdmin):
-    fields = ["class_feature", "hit_dice", "hp_first_level", "hp_higher_levels"]
-    list_display = ["class_feature", "hit_dice", "hp_first_level", "hp_higher_levels"]
+    fields = ["klass_feature", "hit_dice", "hp_first_level", "hp_higher_levels"]
+    list_display = ["klass_feature", "hit_dice", "hp_first_level", "hp_higher_levels"]
 
 
-class ClassAdvancementAdmin(admin.ModelAdmin):
-    fields = ["class_name", "level", "proficiency_bonus"]
-    list_display = ["class_name", "level", "proficiency_bonus"]
-    ordering = ["class_name", "level"]
+class KlassAdvancementAdmin(admin.ModelAdmin):
+    fields = ["klass", "level", "proficiency_bonus"]
+    list_display = ["klass", "level", "proficiency_bonus"]
+    ordering = ["klass", "level"]
 
 
 # Character
@@ -43,4 +43,4 @@ admin.site.register(Sense, SenseAdmin)
 
 # Classes
 admin.site.register(HitPoints, HitPointsAdmin)
-admin.site.register(ClassAdvancement, ClassAdvancementAdmin)
+admin.site.register(KlassAdvancement, KlassAdvancementAdmin)