+++ character/models/character.py
@@ -9,7 +9,7 @@ from utils.dice import Dice
 from ..utils.cache import advancement_key
 from .abilities import Ability
 from .advancement import Advancement
-from .classes import Class
+from .klasses import Klass
 from .equipment import Inventory
 from .races import Alignment, Language, Race, Sense, Size
 from .skills import Skill
@@ -24,9 +24,7 @@ class Character(models.Model):
     name = models.CharField(max_length=100, unique=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     race = models.CharField(max_length=1, choices=Race.choices)
-    class_name = models.CharField(
-        max_length=1, choices=Class.choices, verbose_name="class"
-    )
+    klass = models.CharField(max_length=1, choices=Klass.choices, verbose_name="class")
     level = models.SmallIntegerField(default=1)
     xp = models.IntegerField(default=0)
     hp = models.SmallIntegerField(default=100)