+++ character/models/klasses.py
@@ -3,24 +3,24 @@ from django.db import models
 from .equipment import Equipment
 
 
-class Class(models.TextChoices):
+class Klass(models.TextChoices):
     CLERIC = "C", "Cleric"
     FIGHTER = "F", "Fighter"
     ROGUE = "R", "Rogue"
     WIZARD = "W", "Wizard"
 
 
-class ClassFeature(models.Model):
-    class_name = models.CharField(max_length=1, choices=Class.choices, unique=True)
+class KlassFeature(models.Model):
+    klass = models.CharField(max_length=1, choices=Klass.choices, unique=True)
     equipment = models.ManyToManyField(Equipment)
 
     def __str__(self):
-        return f"{self.get_class_name_display()}'s feature"
+        return f"{self.get_klass_display()}'s feature"
 
 
 class HitPoints(models.Model):
-    class_feature = models.OneToOneField(
-        ClassFeature, on_delete=models.SET_NULL, blank=True, null=True
+    klass_feature = models.OneToOneField(
+        KlassFeature, on_delete=models.SET_NULL, blank=True, null=True
     )
     hit_dice = models.CharField(max_length=5)
     hp_first_level = models.SmallIntegerField()
@@ -30,16 +30,16 @@ class HitPoints(models.Model):
         verbose_name_plural = "hit points"
 
     def __str__(self):
-        return f"{self.class_feature} hit points"
+        return f"{self.klass_feature} hit points"
 
 
-class ClassAdvancement(models.Model):
-    class_name = models.CharField(max_length=1, choices=Class.choices)
+class KlassAdvancement(models.Model):
+    klass = models.CharField(max_length=1, choices=Klass.choices)
     level = models.SmallIntegerField()
     proficiency_bonus = models.SmallIntegerField()
 
     class Meta:
-        ordering = ["class_name"]
+        ordering = ["klass"]
 
     def __str__(self):
         return str(self.level)