+++ character/tests/models/test_classes.py
@@ -1,19 +1,19 @@
 import pytest
 
-from character.models.classes import ClassAdvancement
+from character.models.klasses import KlassAdvancement
 
 
 @pytest.mark.django_db
 class TestClassAdvancementModel:
-    class_advancement = None
+    klass_advancement = None
 
     @pytest.fixture(autouse=True)
     def setup(self):
         # Fixtures are automatically loaded during the test session initialization.
-        self.class_advancement = ClassAdvancement.objects.last()
+        self.klass_advancement = KlassAdvancement.objects.last()
 
     def test_creation(self):
-        assert isinstance(self.class_advancement, ClassAdvancement)
+        assert isinstance(self.klass_advancement, KlassAdvancement)
 
     def test_str(self):
-        assert str(self.class_advancement) == str(self.class_advancement.level)
+        assert str(self.klass_advancement) == str(self.klass_advancement.level)