+++ character/views/equipment.py
@@ -7,7 +7,7 @@ from ..forms.equipment import (
     RogueEquipmentSelectForm,
     WizardEquipmentSelectForm,
 )
-from ..models.classes import Class
+from ..models.klasses import Klass
 from ..models.equipment import Armor, Equipment, Gear, Tool, Weapon
 from .mixins import CharacterContextMixin
 
@@ -19,14 +19,14 @@ class EquipmentSelectView(LoginRequiredMixin, CharacterContextMixin, FormView):
         return self.character.get_absolute_url()
 
     def get_form_class(self):
-        match self.character.class_name:
-            case Class.CLERIC:
+        match self.character.klass:
+            case Klass.CLERIC:
                 form_class = ClericEquipmentSelectForm
-            case Class.FIGHTER:
+            case Klass.FIGHTER:
                 form_class = FighterEquipmentSelectForm
-            case Class.ROGUE:
+            case Klass.ROGUE:
                 form_class = RogueEquipmentSelectForm
-            case Class.WIZARD:
+            case Klass.WIZARD:
                 form_class = WizardEquipmentSelectForm
         return form_class
 
@@ -69,8 +69,8 @@ class EquipmentSelectView(LoginRequiredMixin, CharacterContextMixin, FormView):
             pass
 
         # Some equipment is added without selection, depending on character's class.
-        match self.character.class_name:
-            case Class.CLERIC:
+        match self.character.klass:
+            case Klass.CLERIC:
                 Equipment.objects.create(
                     name=Weapon.Name.CROSSBOW_LIGHT, inventory=self.character.inventory
                 )
@@ -82,7 +82,7 @@ class EquipmentSelectView(LoginRequiredMixin, CharacterContextMixin, FormView):
                     name=Armor.Name.SHIELD, inventory=self.character.inventory
                 )
 
-            case Class.ROGUE:
+            case Klass.ROGUE:
                 Equipment.objects.create(
                     name=Weapon.Name.SHORTBOW, inventory=self.character.inventory
                 )
@@ -102,7 +102,7 @@ class EquipmentSelectView(LoginRequiredMixin, CharacterContextMixin, FormView):
                     name=Tool.Name.THIEVES_TOOLS, inventory=self.character.inventory
                 )
 
-            case Class.WIZARD:
+            case Klass.WIZARD:
                 Equipment.objects.create(
                     name=Gear.Name.SPELLBOOK,
                     inventory=self.character.inventory,